FIX: Rename ext_traits to impl_par_methods
