+++ src/parallel/impl_par_methods.rs
